[FLINK-6137] Activate strict checkstyle for flink-cep
