+++ flink-libraries/flink-cep/src/main/java/org/apache/flink/cep/NonDuplicatingTypeSerializer.java
@@ -35,7 +35,7 @@ import java.util.IdentityHashMap;
  * serialized once. If the same object shall be serialized again, then a reference handle is
  * written instead.
  *
- * Avoiding duplication is achieved by keeping an internal identity hash map. This map contains
+ * <p>Avoiding duplication is achieved by keeping an internal identity hash map. This map contains
  * all serialized objects. To make the serializer work it is important that the same serializer
  * is used for a coherent serialization run. After the serialization has stopped, the identity
  * hash map should be cleared.
@@ -107,8 +107,8 @@ public final class NonDuplicatingTypeSerializer<T> extends TypeSerializer<T> {
 
 	/**
 	 * Serializes the given record.
-	 * <p>
-	 * First a boolean indicating whether a reference handle (true) or the object (false) is
+	 *
+	 * <p>First a boolean indicating whether a reference handle (true) or the object (false) is
 	 * written. Then, either the reference handle or the object is written.
 	 *
 	 * @param record The record to serialize.
@@ -128,8 +128,8 @@ public final class NonDuplicatingTypeSerializer<T> extends TypeSerializer<T> {
 
 	/**
 	 * Deserializes an object from the input view.
-	 * <p>
-	 * First it reads a boolean indicating whether a reference handle or a serialized object
+	 *
+	 * <p>First it reads a boolean indicating whether a reference handle or a serialized object
 	 * follows.
 	 *
 	 * @param source The input view from which to read the data.
@@ -172,7 +172,7 @@ public final class NonDuplicatingTypeSerializer<T> extends TypeSerializer<T> {
 	public boolean equals(Object obj) {
 		if (obj instanceof NonDuplicatingTypeSerializer) {
 			@SuppressWarnings("unchecked")
-			NonDuplicatingTypeSerializer<T> other = (NonDuplicatingTypeSerializer<T>)obj;
+			NonDuplicatingTypeSerializer<T> other = (NonDuplicatingTypeSerializer<T>) obj;
 
 			return (other.canEqual(this) && typeSerializer.equals(other.typeSerializer));
 		} else {
