+++ flink-libraries/flink-cep/src/main/java/org/apache/flink/cep/nfa/State.java
@@ -30,8 +30,8 @@ import java.util.Objects;
 
 /**
  * Represents a state of the {@link NFA}.
- * <p>
- * Each state is identified by a name and a state type. Furthermore, it contains a collection of
+ *
+ * <p>Each state is identified by a name and a state type. Furthermore, it contains a collection of
  * state transitions. The state transitions describe under which conditions it is possible to enter
  * a new state.
  *
@@ -59,7 +59,9 @@ public class State<T> implements Serializable {
 		return stateType == StateType.Final;
 	}
 
-	public boolean isStart() { return stateType == StateType.Start; }
+	public boolean isStart() {
+		return stateType == StateType.Start;
+	}
 
 	public String getName() {
 		return name;
@@ -84,7 +86,7 @@ public class State<T> implements Serializable {
 		addStateTransition(StateTransitionAction.IGNORE, this, condition);
 	}
 
-	public void addIgnore(final State<T> targetState,final IterativeCondition<T> condition) {
+	public void addIgnore(final State<T> targetState, final IterativeCondition<T> condition) {
 		addStateTransition(StateTransitionAction.IGNORE, targetState, condition);
 	}
 
@@ -104,7 +106,7 @@ public class State<T> implements Serializable {
 	public boolean equals(Object obj) {
 		if (obj instanceof State) {
 			@SuppressWarnings("unchecked")
-			State<T> other = (State<T>)obj;
+			State<T> other = (State<T>) obj;
 
 			return name.equals(other.name) &&
 				stateType == other.stateType &&
