+++ flink-libraries/flink-cep/src/main/java/org/apache/flink/cep/pattern/Quantifier.java
@@ -15,11 +15,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.flink.cep.pattern;
 
 import java.util.EnumSet;
 import java.util.Objects;
 
+/**
+ * A quantifier describing the Pattern. There are three main groups of {@link Quantifier}.
+ *
+ * <p><ol>
+ *     <li>Single</li>
+ *     <li>Looping</li>
+ *     <li>Times</li>
+ * </ol>
+ *
+ * <p>Each {@link Pattern} can be optional and have a {@link ConsumingStrategy}. Looping and Times have also additional
+ * inner consuming strategy that is applied between accepted events in the pattern
+ *
+ */
 public class Quantifier {
 
 	private final EnumSet<QuantifierProperty> properties;
@@ -36,15 +50,15 @@ public class Quantifier {
 		this.consumingStrategy = consumingStrategy;
 	}
 
-	public static Quantifier ONE(final ConsumingStrategy consumingStrategy) {
+	public static Quantifier one(final ConsumingStrategy consumingStrategy) {
 		return new Quantifier(consumingStrategy, QuantifierProperty.SINGLE);
 	}
 
-	public static Quantifier ONE_OR_MORE(final ConsumingStrategy consumingStrategy) {
+	public static Quantifier oneOrMore(final ConsumingStrategy consumingStrategy) {
 		return new Quantifier(consumingStrategy, QuantifierProperty.LOOPING);
 	}
 
-	public static Quantifier TIMES(final ConsumingStrategy consumingStrategy) {
+	public static Quantifier times(final ConsumingStrategy consumingStrategy) {
 		return new Quantifier(consumingStrategy, QuantifierProperty.TIMES);
 	}
 
@@ -118,6 +132,9 @@ public class Quantifier {
 		OPTIONAL
 	}
 
+	/**
+	 * Describes strategy which events are matched in this {@link Pattern}. See docs for more info.
+	 */
 	public enum ConsumingStrategy {
 		STRICT,
 		SKIP_TILL_NEXT,
