+++ flink-libraries/flink-cep/src/main/java/org/apache/flink/cep/pattern/conditions/IterativeCondition.java
@@ -65,8 +65,8 @@ public abstract class IterativeCondition<T> implements Function, Serializable {
 
 	/**
 	 * The filter function that evaluates the predicate.
-	 * <p>
-	 * <strong>IMPORTANT:</strong> The system assumes that the function does not
+	 *
+	 * <p><strong>IMPORTANT:</strong> The system assumes that the function does not
 	 * modify the elements on which the predicate is applied. Violating this assumption
 	 * can lead to incorrect results.
 	 *
