+++ flink-libraries/flink-cep/src/test/java/org/apache/flink/cep/operator/CEPMigration11to13Test.java
@@ -35,6 +35,7 @@ import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
 import org.apache.flink.streaming.runtime.tasks.OperatorStateHandles;
 import org.apache.flink.streaming.util.KeyedOneInputStreamOperatorTestHarness;
 import org.apache.flink.streaming.util.OneInputStreamOperatorTestHarness;
+
 import org.junit.Test;
 
 import java.net.URL;
@@ -45,6 +46,9 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+/**
+ * Tests for migration from 1.1.x to 1.3.x.
+ */
 public class CEPMigration11to13Test {
 
 	private static String getResourceFilename(String filename) {
@@ -193,7 +197,7 @@ public class CEPMigration11to13Test {
 
 		final Event startEvent = new Event(42, "start", 1.0);
 		final SubEvent middleEvent = new SubEvent(42, "foo", 1.0, 10.0);
-		final Event endEvent=  new Event(42, "end", 1.0);
+		final Event endEvent = new Event(42, "end", 1.0);
 
 		// uncomment these lines for regenerating the snapshot on Flink 1.1
 		/*
